8254999: Move G1RemSetSampligTask to more appropriate location
