+++ src/hotspot/share/gc/g1/g1RemSet.hpp
@@ -49,6 +49,8 @@ class G1ParScanThreadState;
 class G1ParScanThreadStateSet;
 class G1Policy;
 class G1ScanCardClosure;
+class G1ServiceTask;
+class G1ServiceThread;
 class HeapRegionClaimer;
 
 // A G1RemSet in which each heap region has a rem set that records the
@@ -65,6 +67,7 @@ private:
   G1CardTable*           _ct;
   G1Policy*              _g1p;
   G1HotCardCache*        _hot_card_cache;
+  G1ServiceTask*         _sampling_task;
 
   void print_merge_heap_roots_stats();
 
@@ -80,6 +83,9 @@ public:
            G1CardTable* ct,
            G1HotCardCache* hot_card_cache);
   ~G1RemSet();
+
+  // Initialize and schedule young remembered set sampling task.
+  void initialize_sampling_task(G1ServiceThread* thread);
 
   // Scan all cards in the non-collection set regions that potentially contain
   // references into the current whole collection set.
