+++ src/hotspot/share/gc/g1/g1ServiceThread.hpp
@@ -29,7 +29,6 @@
 #include "runtime/mutex.hpp"
 
 class G1PeriodicGCTask;
-class G1RemSetSamplingTask;
 class G1ServiceTaskQueue;
 class G1ServiceThread;
 
@@ -106,7 +105,6 @@ class G1ServiceThread: public ConcurrentGCThread {
   Monitor _monitor;
   G1ServiceTaskQueue _task_queue;
 
-  G1RemSetSamplingTask* _remset_task;
   G1PeriodicGCTask* _periodic_gc_task;
 
   double _vtime_accum;  // Accumulated virtual time.
